+++ src/qts/__init__.py
@@ -23,7 +23,6 @@ from qts._errors import (
     InvalidWrapperError,
     MultipleWrappersAvailableError,
     NoWrapperAvailableError,
-    NoWrapperSelectedError,
     QtsError,
     WrapperAlreadySelectedError,
 )